requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.2.15
django-environ==0.9.0
djangorestframework==3.13.1
idna==3.4
pexpect==4.8.0
ptyprocess==0.7.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
typing-extensions==4.3.0
urllib3==1.26.12
